+++ src/core/elementalEvents.js
@@ -1,0 +1,17 @@
+/**
+ * Event for when an elemental is destroyed
+ * @type {string}
+ */
+export var ELEMENTAL_DESTROYED_EVENT = "instance destroyed";
+
+/**
+ * Event for when an elemental is paused
+ * @type {string}
+ */
+export var ELEMENTAL_PAUSED_EVENT = "instance paused";
+
+/**
+ * Event for when an elemental is resumed
+ * @type {string}
+ */
+export var ELEMENTAL_RESUMED_EVENT = "instance resumed";
